# services/banking-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing accounts, balances, and banking operations.",
        "title": "Banking Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/accounts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of all accounts. Employee access only. Supports filtering by client, account type, kind, status and currency.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List all accounts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by client ID",
                        "name": "client_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by account type",
                        "name": "account_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by account kind",
                        "name": "account_kind",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by currency ID",
                        "name": "currency_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new bank account for a client. Account kind must be \"Current\" or \"Foreign\". Current accounts require a subtype; foreign accounts require a currency code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create a new bank account",
                "parameters": [
                    {
                        "description": "Account creation data",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
        "/api/clients/{clientId}/accounts/{accountNumber}/limits": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Confirms a pending limit change request using the verification code from the mobile app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Confirm account limit change",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "accountNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Verification code generated in mobile app",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConfirmLimitsChangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/clients/{clientId}/accounts/{accountNumber}/limits/request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Initiates a limit change request for an account. Confirmation uses TOTP code generated in the mobile app.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Request account limit change",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "accountNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New daily and monthly limits",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RequestLimitsChangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/clients/{clientId}/accounts/{accountNumber}/name": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the display name of an account owned by the authenticated client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Update account name",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "accountNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New account name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAccountNameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AccountResponse": {
            "type": "object",
            "properties": {
                "account_kind": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "available_balance": {
                    "type": "number"
                },
                "balance": {
                    "type": "number"
                },
                "client_id": {
                    "type": "integer"
                },
                "company_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/model.CurrencyCode"
                },
                "daily_limit": {
                    "type": "number"
                },
                "daily_spending": {
                    "type": "number"
                },
                "employee_id": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "string"
                },
                "maintenance_fee": {
                    "type": "number"
                },
                "monthly_limit": {
                    "type": "number"
                },
                "monthly_spending": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "reserved_funds": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "subtype": {
                    "type": "string"
                }
            }
        },
...
        "dto.ConfirmLimitsChangeRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "dto.CreateAccountRequest": {
            "type": "object",
            "required": [
                "account_kind",
                "account_type",
                "client_id",
                "employee_id",
                "expires_at",
                "name"
            ],
            "properties": {
                "account_kind": {
                    "$ref": "#/definitions/model.AccountKind"
                },
                "account_type": {
                    "$ref": "#/definitions/model.AccountType"
                },
                "client_id": {
                    "type": "integer"
                },
                "company_id": {
                    "type": "integer"
                },
                "create_card": {
                    "type": "boolean"
                },
                "currency_code": {
                    "description": "required if AccountKind=Foreign",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.CurrencyCode"
                        }
                    ]
                },
                "employee_id": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "string"
                },
                "generate_card": {
                    "type": "boolean"
                },
                "initial_balance": {
                    "type": "number",
                    "minimum": 0
                },
                "name": {
                    "type": "string"
                },
                "subtype": {
                    "description": "required if AccountKind=Current",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Subtype"
                        }
                    ]
                }
            }
        },
...
        "dto.RequestLimitsChangeRequest": {
            "type": "object",
            "required": [
                "daily_limit",
                "monthly_limit"
            ],
            "properties": {
                "daily_limit": {
                    "type": "number"
                },
                "monthly_limit": {
                    "type": "number"
                }
            }
        },
...
        "dto.UpdateAccountNameRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.AccountKind": {
            "type": "string",
            "enum": [
                "Current",
                "Foreign",
                "Internal"
            ],
            "x-enum-varnames": [
                "AccountKindCurrent",
                "AccountKindForeign",
                "AccountKindInternal"
            ]
        },
        "model.AccountType": {
            "type": "string",
            "enum": [
                "Personal",
                "Business",
                "Bank"
            ],
            "x-enum-varnames": [
                "AccountTypePersonal",
                "AccountTypeBusiness",
                "AccountTypeBank"
            ]
        },
        "model.CurrencyCode": {
            "type": "string",
            "enum": [
                "EUR",
                "USD",
                "CHF",
                "GBP",
                "JPY",
                "CAD",
                "AUD",
                "RSD"
            ],
            "x-enum-varnames": [
                "EUR",
                "USD",
                "CHF",
                "GBP",
                "JPY",
                "CAD",
                "AUD",
                "RSD"
            ]
        },
...
        "model.Subtype": {
            "type": "string",
            "enum": [
                "Standard",
                "Savings",
                "Pension",
                "Youth",
                "Student",
                "Unemployed",
                "LLC",
                "JointStock",
                "Foundation"
            ],
            "x-enum-varnames": [
                "SubtypeStandard",
                "SubtypeSavings",
                "SubtypePension",
                "SubtypeYouth",
                "SubtypeStudent",
                "SubtypeUnemployed",
                "SubtypeLLC",
                "SubtypeJointStock",
                "SubtypeFoundation"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}